refactor(actions): tasks/ como directorio con barril
