+++ src/actions/tasks/createTask.ts
@@ -1,0 +1,19 @@
+'use server'
+
+import { prisma } from '@/lib/prisma'
+import { requireColumnAccess } from '../shared'
+
+export async function createTask({
+	columnId,
+	descriptionText,
+}: {
+	columnId: string
+	descriptionText: string
+}): Promise<string> {
+	await requireColumnAccess(columnId)
+
+	const task = await prisma.task.create({
+		data: { descriptionText, columnId },
+	})
+	return task.id
+}
