+++ src/actions/tasks/deleteColumn.ts
@@ -1,0 +1,9 @@
+'use server'
+
+import { prisma } from '@/lib/prisma'
+import { requireColumnAccess } from '../shared'
+
+export async function deleteColumn({ columnId }: { columnId: string }): Promise<void> {
+	await requireColumnAccess(columnId)
+	await prisma.column.delete({ where: { id: columnId } })
+}
