+++ src/actions/tasks/deleteTask.ts
@@ -1,0 +1,9 @@
+'use server'
+
+import { prisma } from '@/lib/prisma'
+import { requireTaskAccess } from '../shared'
+
+export async function deleteTask({ taskId }: { taskId: string }): Promise<void> {
+	await requireTaskAccess(taskId)
+	await prisma.task.delete({ where: { id: taskId } })
+}
