+++ src/actions/tasks/updateTask.ts
@@ -1,0 +1,30 @@
+'use server'
+
+import { prisma } from '@/lib/prisma'
+import { requireTaskAccess } from '../shared'
+
+export async function updateTask({
+	taskId,
+	descriptionText,
+	notesAndComments,
+	tags,
+	timelineHistory,
+}: {
+	taskId: string
+	descriptionText?: string
+	notesAndComments?: string
+	tags?: object
+	timelineHistory?: object
+}): Promise<void> {
+	await requireTaskAccess(taskId)
+
+	await prisma.task.update({
+		where: { id: taskId },
+		data: {
+			...(descriptionText !== undefined && { descriptionText }),
+			...(notesAndComments !== undefined && { notesAndComments }),
+			...(tags !== undefined && { tags }),
+			...(timelineHistory !== undefined && { timelineHistory }),
+		},
+	})
+}
